👷‍♂️ Refactors the code     - Adds 3 new components         - BottomSheet         - LoadingAnimation         - MainButton 👷‍♂️ Reduces length of App.js 👷‍♂️ Adds screenshots of app for README
